+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 ## lifebit-ai/cloudos-cli: changelog
 
+## v2.39.0 (2025-07-09)
+
+### Feat
+
+- Adds `job clone` and `job resume` commands
+- Refactors `job run` command
+
+## v2.38.3 (2025-07-08)
+
+### Fix
+
+- Array file project
+
 ## v2.38.2 (2025-07-07)
 
 ### Fix